breaks out icon specific config
